Full network stack implementation (-tests) - Removed test activity - Ensured all DTO's are up to date - Refactored networking, added basic doc - Added user details card

+++ app/src/main/java/com/ig/igtradinggame/ui/cards/balance/BalanceCard.java
@@ -30,7 +30,7 @@ public final class BalanceCard extends BaseCardView {
     public void setup(CardModel cardModel) {
         if (cardModel instanceof BalanceModel) {
             BalanceModel viewModel = (BalanceModel) cardModel;
-            Integer balance = viewModel.getBalance();
+            Double balance = viewModel.getBalance();
             balanceText.setText(balance.toString());
         }
     }
